+++ src/ace_segment/ht16k33/Ht16k33Module.h
@@ -125,7 +125,7 @@ class Ht16k33Module : public LedModule {
       mPatterns[pos] = pattern;
     }
 
-    uint8_t getPatternAt(uint8_t pos) override {
+    uint8_t getPatternAt(uint8_t pos) const override {
       return mPatterns[pos];
     }