+++ src/ace_segment/max7219/Max7219Module.h
@@ -153,7 +153,7 @@ class Max7219Module : public LedModule {
       mPatterns[pos] = pattern;
     }
 
-    uint8_t getPatternAt(uint8_t pos) override {
+    uint8_t getPatternAt(uint8_t pos) const override {
       return mPatterns[pos];
     }