+++ src/ace_segment/scanning/ScanningModule.h
@@ -138,7 +138,7 @@ class ScanningModule : public LedModule {
       mPatterns[pos] = pattern;
     }
 
-    uint8_t getPatternAt(uint8_t pos) override {
+    uint8_t getPatternAt(uint8_t pos) const override {
       return mPatterns[pos];
     }