+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 * Unreleased
+    * Make `LedModule::getPatternAt()` into a `const` function.
 * 0.8.1 (2021-07-31)
     * Add `HelloTm1636`, `HelloMax7219`, `HelloHt16k33`, and `HelloHc595`
       examples to `examples/` and `README.md`.